refactor: (JDS) Textarea 서브컴포넌트를 compound/ 폴더로 분리
- 내부 Field 컴포넌트와 동일하게 Content/Control/Counter를 compound/ 폴더의 개별 파일로 분리하고, root(Textarea.tsx)는 Object.assign으로 조립.
- Counter는 context 자동 추적을 유지(길이는 TextareaProvider에서 관리)하고 counter 스타일을 단순화.
- Figma 기준으로 text content min-height 48px, counter가 있을 때 textarea↔counter gap 12px 적용.
