+++ packages/jds/src/components/Textarea/index.ts
@@ -1,5 +1,6 @@
 export { Textarea } from "./Textarea";
-export type { TextareaProps, TextareaControlProps } from "./Textarea";
+export type { TextareaProps } from "./Textarea";
+export type { TextareaControlProps } from "./compound/Control";
 export type {
   FieldContentProps as TextareaContentProps,
   FieldLabelProps as TextareaLabelProps,
